Изменение типов данных с int на long

+++ app/src/main/java/ru/sis/statube/db/entity/DataDailyEntity.kt
@@ -8,8 +8,8 @@ open class DataDailyEntity : RealmObject() {
 
     open var date: Long = 0
 
-    open var subs: Int? = null
+    open var subs: Long? = null
 
-    open var views: Int? = null
+    open var views: Long? = null
 
 }
