+++ app/src/main/java/ru/sis/statube/db/entity/ValueByPeriodEntity.kt
@@ -6,16 +6,16 @@ import io.realm.annotations.RealmClass
 @RealmClass
 open class ValueByPeriodEntity : RealmObject() {
 
-    open var by14: Int? = null
+    open var by14: Long? = null
 
-    open var by30: Int? = null
+    open var by30: Long? = null
 
-    open var by60: Int? = null
+    open var by60: Long? = null
 
-    open var by90: Int? = null
+    open var by90: Long? = null
 
-    open var by180: Int? = null
+    open var by180: Long? = null
 
-    open var by365: Int? = null
+    open var by365: Long? = null
 
 }
