+++ app/src/main/java/ru/sis/statube/model/ValueByPeriod.kt
@@ -4,16 +4,16 @@ import java.io.Serializable
 
 class ValueByPeriod : Serializable {
 
-    var by14: Int? = null
+    var by14: Long? = null
 
-    var by30: Int? = null
+    var by30: Long? = null
 
-    var by60: Int? = null
+    var by60: Long? = null
 
-    var by90: Int? = null
+    var by90: Long? = null
 
-    var by180: Int? = null
+    var by180: Long? = null
 
-    var by365: Int? = null
+    var by365: Long? = null
 
 }
